% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs_equation.R
\name{pcs_equation}
\alias{pcs_equation}
\title{PCS calculation}
\usage{
pcs_equation(corr, t = 100, d = 0.05, min = 0, max = 1, in_act = 0.1)
}
\arguments{
\item{corr}{Dataframe of a correlation matrix}

\item{t}{Maximum number of iterations, standard 100}

\item{d}{Decay parameter, standard .05}

\item{min}{Minimum activation, standard 0}

\item{max}{Maximum activation, standard 1}

\item{in_act}{Initial activation, standard .1}
}
\value{
Returns activation of each node each iteration in dataframe with the timepoint.
}
\description{
This function takes a correlation matrix (as data frame!) as input and returns the values of a PCS model.
In an iterative process, nodes are either activated or inhibited to represent a mental model.
}
\examples{
pcs_equation(data.frame(X = c(1, .8, .2), Y = c(.8, 1, -.5), Z = c(.2, -.5, 1)))

}
