% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs_viz.R
\name{pcs_viz}
\alias{pcs_viz}
\title{PCS Visualization}
\usage{
pcs_viz(pcsresult)
}
\arguments{
\item{pcsresult}{Dataframe of activation as produced in the function \emph{pcs_equation}.}
}
\value{
Returns a line plot showing the activation progress of the input nodes.
}
\description{
This function takes a dataframe of PCS activations over a number of iterations as input and
returns a plot of active nodes in a green-blue color and inactive nodes in a red-pink color.
}
\examples{
pcs_viz(pcs_equation(data.frame(X = c(1, .8, .2), Y = c(.8, 1, -.5), Z = c(.2, -.5, 1))))

}
